+++ lib/rspec/change_to_now/matchers.rb
@@ -1,5 +1,6 @@
 require 'rspec/change_to_now/matchers/detect'
 require 'rspec/change_to_now/matchers/negate'
 require 'rspec/change_to_now/matchers/matcher_only'
+require 'rspec/change_to_now/matchers/as_matcher'
 require 'rspec/change_to_now/matchers/dsl'